gcp-document-apis/functions/backup_manager/requirements.txt
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-logging==3.8.0
